chore: add commitlint and use conventional config
Enforce commit messages are standardised.  I am trying out setting this
up as the first thing in the project rather than later to ensure that
all commits follow this convention and not just the ones in the future.

The conventional configuration seems to be adopted by the community as
a good standard to start with.